requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
alembic==1.13.1
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio[trio]==3.7.1
    # via
    #   fastapi-login
    #   httpx
    #   starlette
    #   watchfiles
async-generator==1.10
    # via trio
attrs==23.2.0
    # via
    #   outcome
    #   trio
bcrypt==3.2.0
    # via
    #   -r requirements.in
    #   passlib
certifi==2024.6.2
    # via
    #   httpcore
    #   httpx
cffi==1.16.0
    # via
    #   bcrypt
    #   trio
click==8.1.7
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
dnspython==2.6.1
    # via email-validator
email-validator==2.2.0
    # via
    #   -r requirements.in
    #   fastapi
fastapi==0.111.0
    # via
    #   -r requirements.in
    #   fastapi-login
fastapi-cli==0.0.4
    # via fastapi
fastapi-login==1.10.0
    # via -r requirements.in
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via fastapi
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   trio
jinja2==3.1.4
    # via
    #   -r requirements.in
    #   fastapi
mako==1.3.5
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
orjson==3.10.5
    # via fastapi
outcome==1.3.0.post0
    # via trio
passlib[bcrypt]==1.7.4
    # via -r requirements.in
pycparser==2.22
    # via cffi
pydantic==2.7.4
    # via fastapi
pydantic-core==2.18.4
    # via pydantic
pygments==2.18.0
    # via rich
pyjwt==2.8.0
    # via fastapi-login
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
python-multipart==0.0.9
    # via
    #   -r requirements.in
    #   fastapi
pyyaml==6.0.1
    # via uvicorn
rich==13.7.1
    # via typer
shellingham==1.5.4
    # via typer
six==1.16.0
    # via bcrypt
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   trio
sortedcontainers==2.4.0
    # via trio
sqlalchemy==2.0.31
    # via
    #   -r requirements.in
    #   alembic
starlette==0.37.2
    # via fastapi
trio==0.21.0
    # via anyio
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   alembic
    #   fastapi
    #   fastapi-login
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
ujson==5.10.0
    # via fastapi
uvicorn[standard]==0.30.1
    # via
    #   -r requirements.in
    #   fastapi
watchfiles==0.22.0
    # via uvicorn
websockets==12.0
    # via uvicorn
